more get_collections to run_ml_model to be more easily able to change which collections shall be used.
